refactor(gateway): split the messages translation layer by concern
Addresses CodeRabbit review on #6207.

Split anthropic_translation.go, which had grown past the 300-source-line
limit, into four files along its existing seams: error envelopes, the
framing state machine, the stream iterator, and request building plus
dispatch. Each is now well under the limit.

Also from the review: name 529 as overloaded_error so an Anthropic
overload relayed from upstream stays retryable rather than terminal;
drop a comment that promised a local stream close the cancellation path
never performs; guard the test helpers' index dereferences and the
sequence-length indexing so a framing regression fails as an assertion
instead of panicking; and lock the httptest handlers' captured bodies,
which the test goroutine read without synchronisation.
